feat: retrieval API file scanning verdict endpoint
Update the Form retrieval API to add an endpoint that accepts a
form ID and submission ID and returns the file attachments
for the given submission, along with their file scan verdicts.

The endpoint introduces two new middleware:

1. `validateQueryParams`: checks that a list of named
query parameters are valid according to a provided
`isValid` function.
2. `validFormSubmission`: checks that a given form ID
and submission ID represent an existing form submission
in the Vault.  The `FormSubmission` JSON string is then
returned to the endpoint handler via a new
`MiddlewareProps.formSubmission`.