Adding a Modal component
The markup is Inspired by the bootstrap modal and also the react-modal library,
and the actual implementation was helped along by a helpful turorial.

The modal opens and closes as you would expect by clicking the buttons,
and also closes if you press the escape key or click outside of the modal.

Also added more nicely styled buttons to more or less match what we have in Figma.

Sources:
- https://github.com/reactjs/react-modal
- https://getbootstrap.com/docs/4.0/components/modal/
- https://medium.com/tinyso/how-to-create-a-modal-component-in-react-from-basic-to-advanced-a3357a2a716a