Don't replace the entire "form", just reset the elements
There was a bug I was seeing when either moving form fields or
removing a field, which was that a bunch of the fields in the "form"
object were being removed.

The way that the new "elements" was being updated was correct, but
it was resetting the "form" object such that all the other keys
were being deleted.

Updated so that the "elements" key is getting replaced without
any side effects.