fix: 'remove' buttons are larger and add blue focus states
Added styles from the CDS design system to our buttons and the
radios and checkboxes. Basically, it's this really bright blue color
and it puts us more in line with the design system.

The biggest change though is the 'remove' buttons, which now are
much easier to see and click (larger hit area) than before.