fix: generalize the code for inputs and add focus styles to them
We need inputs in a few files, so I created an Input.tsx file that's
really basic and that way we can share styles around the app.

Brought in the same blue focus style the design system uses and that
we already have for buttons and checkboxes.

Also updated the question title to have just an underline, a bolder
weight, and no focus state.