Update form builder so it looks more like the Figma mockups (#1039)
* feat: Add GDS/CDS checkboxes & radios

Since we are adding some to the modal and we already have one
on the questions panel, I figured it was time to add the CDS checkboxes
and radios in.

Changed the size of the checkbox a bit and made sure the check itself
was centered.

Source:
- https://design-system.service.gov.uk/components/checkboxes/

* feat: Add option by default to new questions

This is something Carolyn pointed out, if we could add a new, empty
option each time we create a new question.
Added it to the defaul field definition in the templateStore.

Also added an effect that will focus the question title by default
instead of the option.

* feat: Add 'required' checkbox to modal

It's in the mockups, so let's bring it into the modal.

* fix: 'remove' buttons are larger and add blue focus states

Added styles from the CDS design system to our buttons and the
radios and checkboxes. Basically, it's this really bright blue color
and it puts us more in line with the design system.

The biggest change though is the 'remove' buttons, which now are
much easier to see and click (larger hit area) than before.

* fix: shrink the 'options'

Basically, all the text in the question panel should be smaller,
with the one exception being the question itself.

* fix: generalize the code for inputs and add focus styles to them

We need inputs in a few files, so I created an Input.tsx file that's
really basic and that way we can share styles around the app.

Brought in the same blue focus style the design system uses and that
we already have for buttons and checkboxes.

Also updated the question title to have just an underline, a bolder
weight, and no focus state.

* fix: add new options by pressing 'enter'

When you press "enter" from an option input, it should add a new
input.

Always a bit risky adding hotkeys but I think we are safe here.