test: add test for the unset method on the ModalStore
In the past commit, I updated the ModalStore to include an "unset"
method (same as TemplateStore) because once we start dealing with
subtypes of the text field or the maxLength stuff, then we have to
be unsetting those when we change field types, and with dict
comprehensions it's way more complex and harder to tell what's going on.

Added some tests here to make sure unset works as it should.