Only focus inputs when we want them to be focused
Only after adding an element or adding an option do we want to add
focus to those inputs.

This commit adds a value to the store that keeps track of whether or
not we want to autofocus a question title or an option input.

This way, we can switch between tabs without the focus shooting down the page.