feat: Double-click headers to switch language
We need to get the French/English language toggle out of the top
nav but we also want it to be available to us.

The compromise here is using the Headings as the toggle, even though
it is not something we want to do long-term.

Still easy to click and very unlikely someone will activate it by accident.