Import "textField" elements with a validation type of "text"
There are a handful of accepted validation "types" in the JSON schema.

Here they are:
- ["email", "alphanumeric", "text", "name", "number", "phone", "date"]

The "text" one I hadn't cottoned onto before, because it seems like it
means there are no special validation rules.

Our logic thus far assumes that "textField" elements with validation
rules should be treated differently (like they are a date or an email),
whereas "text" never really showed up as an option.

Recently, while importing an already-existing JSON form, I noticed that the
text elements were being shown as checkboxes, and this was the root cause.

Updated the logic so that a "textField" with no validation type and a "textField"
with a validation type of "text" are treated the same.