fix: Update design for form title and rich text fields
Our rich text fields were full-width and the form title looked like
a regular text input.

This PR makes the rich text fields use the same horizontal spacing
as the other elements, and it changes the form title to be the one underline.

Also: resize the SVG for the start page since it was too small.