Replace "Add element" buttons and modal buttons with tailwind button
We want to move off styled components, so this is replaces all the
"Add element" buttons in the form builder.

It also replaces the "save" and "cancel" buttons at the bottom of
the modal.

It does not change the buttons in the PanelActions component or the
"icon buttons" that we are using for removing an option.