Use getter for focusInput rather than value
Tracing through the code, we found out that 'focusInput' was not
getting updated between components. It would be "true" for subsequent
components, even if a previous one called "setFocusInput(false)".

I added a get function to the store so that we can just return the
current value of focusInput in the useEffect hook itself, rather than
getting one value and that value not updating anymore.