Add a skip link to the form builder
Adding a skip link to the form builder that skips our 3 levels of
navigation and shoots you down straight into the normal page content.

It's a super basic component that takes no properties and expects
"#content" to be on the page somewhere.

Note: I didn't use the translate function here on purpose because I
figure this component might get used elsewhere and it didn't seem like
we want to introduce the whole namespace thing. It's easy enough
to add these strings to the locales files, but it is also totally
self contained right now.

Copy is taken from Canada.ca, so it should be pretty reliable.

Used this tutorial for the basic approach:

- https://levelup.gitconnected.com/build-an-accessible-skip-to-content-anchor-link-with-react-140903f3bd7e