Add question number to label
Previously, the label was always "Question" but now it includes the
number of the question as well, which is pretty innovative.

Also, for 2 digit numbers, the grey background overlaps with the
question text input, so I put in a CSS fix to change that.

Note that 3 digit numbers will still overlap but it can't be helped.