feat: add Dockerfile to creat Lambda PR review env
Add a Dockerfile that can run the form viewer client as a Docker
based Lambda function.  This is essentially the same Dockerfile
as the ECS task Dockerfile, with the addition of the AWS Lambda
web adapter:

https://github.com/awslabs/aws-lambda-web-adapter

It also uses a different yarn `start:pr` script that does not perform
any database migrations.  This is done to prevent a PR review
environment from triggering database changes before they have
been reviewed and approved.