feat: workflow to sync PR review env vars (#2282)
Add a workflow that synchronizes the latest ECS task's env vars
and secrets with the ParameterStore used to set these values
on the PR review lambda functions.

This allows the latest ECS task deployed to Staging to act as the
source of truth for env vars and secrets.  It also means we do
not need to redefine all these values for the PR review functions.