fix: allow generate SBOM step to fail (#4348)
Update the Staging ECR push workflow to allow the generate SBOM step
to fail.  This step is intermittently failing because the public Trivy Docker image
that contains the vulnerability database is being rate limited.

SRE is working on setting up a dedicated ECR for CDS projects to fix
this rate limiting in the future.